Added additional KEY_OTHER values to fix more warnings.

+++ src/ActorMultiVertex.cpp
@@ -389,7 +389,7 @@ CubicSplineN* ActorMultiVertex::GetSpline(size_t i)
 	return &(_splines[i]);
 }
 
-void ActorMultiVertex::SetState(size_t i)
+void ActorMultiVertex::SetState(int i)
 {
 	ASSERT(i < _states.size());
 	_cur_state= i;
